fix null problem api

+++ Brief13-STC-API/Br13/src/main/java/com/project/Br13/controller/ActiviteController.java
@@ -23,7 +23,7 @@ public class ActiviteController {
 	ActiviteServices activiteservices;
 	
 	@GetMapping("/{id}")
-	public Activite getActivite(@PathVariable Integer id_activite) {
+	public Activite getActivite(@PathVariable Long id_activite) {
 	Activite activite = activiteservices.getById(id_activite);
 	return activite;
 	}
@@ -40,7 +40,7 @@ public class ActiviteController {
 
 
 	@DeleteMapping("/{id}")
-	public void DeleteActivite(@PathVariable Integer id_activite) {
+	public void DeleteActivite(@PathVariable Long id_activite) {
 		activiteservices.delete(id_activite);
 	}
 }
